feat(osinfo): separate PDF reconstruction flows into dedicated pipeline
Separates PDF reconstruction functionality from rj_cvl__osinfo_mongo into
a new dedicated pipeline rj_cvl__osinfo_pdf_reconstruct for better deployment
management and follows repository best practices.

Changes:
- Create new pipeline rj_cvl__osinfo_pdf_reconstruct with unified flow supporting
  both individual and batch PDF reconstruction modes
- Implement atomic tasks following repository patterns (no task-to-task calls)
- Configure default GCS storage: gs://rj-iplanrio/staging/brutos_osinfo_mongo/files_pdfs
- Remove PDF reconstruction flows and tasks from rj_cvl__osinfo_mongo
- Clean up deployments: separate staging/prod for each pipeline
