+++ pipelines/rj_cvl__osinfo_mongo/flow.py
@@ -28,11 +28,7 @@ from .tasks import (
     build_batch_query,
     chunk_list,
     get_batch_dump_mode,
-    get_chunks_from_bigquery,
     get_files_ids_from_bigquery,
-    reconstruct_pdf,
-    save_pdf_to_gcs,
-    validate_pdf,
 )
 
 
@@ -158,99 +154,3 @@ def rj_cvl__osinfo_mongo(
             only_staging_dataset=only_staging_dataset,
             add_timestamp_column=add_timestamp_column,
         )
-
-
-@flow(log_prints=True)
-def rj_cvl__osinfo_pdf_reconstruct(
-    files_ids: list[str],
-    dataset_id: str = "brutos_osinfo_mongo",
-    table_id: str = "chunks",
-    gcs_bucket: Optional[str] = None,
-    gcs_prefix: str = "reconstructed_pdfs",
-    validate_only: bool = False,
-):
-    """
-    Reconstruct PDFs from BigQuery chunks
-
-    Args:
-        files_ids: List of files_id to reconstruct
-        dataset_id: BigQuery dataset containing chunks
-        table_id: BigQuery table name (default: chunks)
-        gcs_bucket: GCS bucket to save PDFs (optional, if None only validates)
-        gcs_prefix: Prefix/folder in GCS bucket
-        validate_only: If True, only validate without saving to GCS
-
-    Example:
-        files_ids = ["673489d2c4e5f5fa65e31852", "673489d2c4e5f5fa65e31853"]
-        rj_cvl__osinfo_pdf_reconstruct(files_ids, validate_only=True)
-    """
-    chunks_by_file = get_chunks_from_bigquery(
-        files_ids=files_ids,
-        dataset_id=dataset_id,
-        table_id=table_id
-    )
-
-    for file_id, chunks in chunks_by_file.items():
-        pdf_data = reconstruct_pdf(file_id, chunks)
-        validation = validate_pdf(file_id, pdf_data)
-
-        if not validate_only and validation['is_valid'] and gcs_bucket:
-            save_pdf_to_gcs(
-                file_id=file_id,
-                pdf_data=pdf_data,
-                bucket_name=gcs_bucket,
-                prefix=gcs_prefix
-            )
-
-
-@flow(log_prints=True)
-def rj_cvl__osinfo_pdf_reconstruct_batch(
-    bq_files_ids_query: str,
-    dataset_id: str = "brutos_osinfo_mongo",
-    table_id: str = "chunks",
-    gcs_bucket: Optional[str] = None,
-    gcs_prefix: str = "reconstructed_pdfs",
-    batch_size: int = 100,
-    max_files: Optional[int] = None,
-):
-    """
-    Reconstruct PDFs in batches using BigQuery query to get files_ids
-
-    Args:
-        bq_files_ids_query: SQL query to get files_id list
-        dataset_id: BigQuery dataset containing chunks
-        table_id: BigQuery table name
-        gcs_bucket: GCS bucket to save PDFs
-        gcs_prefix: Prefix/folder in GCS bucket
-        batch_size: Number of files to process per batch
-        max_files: Maximum number of files to process (None = all)
-
-    Example:
-        bq_files_ids_query = "SELECT DISTINCT files_id FROM `rj-cvl.brutos_osinfo_mongo.chunks` LIMIT 100"
-        rj_cvl__osinfo_pdf_reconstruct_batch(bq_files_ids_query, gcs_bucket="rj-cvl-osinfo", batch_size=10)
-    """
-    files_ids = get_files_ids_from_bigquery(bq_files_ids_query)
-
-    if max_files:
-        files_ids = files_ids[:max_files]
-
-    batches = chunk_list(files_ids, batch_size)
-
-    for batch in batches:
-        chunks_by_file = get_chunks_from_bigquery(
-            files_ids=batch,
-            dataset_id=dataset_id,
-            table_id=table_id
-        )
-
-        for file_id, chunks in chunks_by_file.items():
-            pdf_data = reconstruct_pdf(file_id, chunks)
-            validation = validate_pdf(file_id, pdf_data)
-
-            if validation['is_valid'] and gcs_bucket:
-                save_pdf_to_gcs(
-                    file_id=file_id,
-                    pdf_data=pdf_data,
-                    bucket_name=gcs_bucket,
-                    prefix=gcs_prefix
-                )
