+++ pipelines/rj_cvl__osinfo_pdf_reconstruct/Dockerfile
@@ -1,0 +1,11 @@
+FROM ghcr.io/prefeitura-rio/prefect_rj_iplanrio:latest
+
+LABEL org.opencontainers.image.source=https://github.com/prefeitura-rio/prefect_rj_iplanrio
+
+WORKDIR /opt/prefect/prefect_rj_iplanrio
+
+COPY ./pyproject.toml ./uv.lock /opt/prefect/prefect_rj_iplanrio/
+
+COPY ./pipelines/rj_cvl__osinfo_pdf_reconstruct ./pipelines/rj_cvl__osinfo_pdf_reconstruct/
+
+RUN uv sync --all-packages
