+++ pipelines/rj_cvl__osinfo_pdf_reconstruct/__init__.py
@@ -1,0 +1,4 @@
+# -*- coding: utf-8 -*-
+"""
+Pipeline para reconstrução de arquivos PDFs do OSINFO (MongoDB)
+"""
