+++ pipelines/rj_cvl__osinfo_pdf_reconstruct/prefect.yaml
@@ -1,4 +1,4 @@
-name: rj_cvl__osinfo_mongo
+name: rj_cvl__osinfo_pdf_reconstruct
 prefect-version: 3.4.3
 
 build:
@@ -10,8 +10,8 @@ build:
       id: build-image
       requires: prefect-docker>=0.6.5
       image_name: ghcr.io/prefeitura-rio/prefect_rj_iplanrio/deployments
-      tag: "rj_cvl__osinfo_mongo-{{ get-commit-hash.stdout }}"
-      dockerfile: pipelines/rj_cvl__osinfo_mongo/Dockerfile
+      tag: "rj_cvl__osinfo_pdf_reconstruct-{{ get-commit-hash.stdout }}"
+      dockerfile: pipelines/rj_cvl__osinfo_pdf_reconstruct/Dockerfile
 
 push:
   - prefect_docker.deployments.steps.push_docker_image:
@@ -24,25 +24,25 @@ pull:
       directory: /opt/prefect/prefect_rj_iplanrio
 
 deployments:
-  - name: rj-cvl--osinfo-mongo--staging
+  - name: rj-cvl--osinfo-pdf-reconstruct--staging
     version: "{{ get-commit-hash.stdout }}"
-    entrypoint: pipelines/rj_cvl__osinfo_mongo/flow.py:rj_cvl__osinfo_mongo
+    entrypoint: pipelines/rj_cvl__osinfo_pdf_reconstruct/flow.py:rj_cvl__osinfo_pdf_reconstruct
     work_pool:
       name: k3s-pool
       work_queue_name: default
       job_variables:
         image: "{{ build-image.image_name }}:{{ build-image.tag }}"
-        command: uv run --package rj_cvl__osinfo_mongo -- prefect flow-run execute
+        command: uv run --package rj_cvl__osinfo_pdf_reconstruct -- prefect flow-run execute
         secretName: prefect-jobs-secrets
         image_pull_policy: Always
-  - name: rj-cvl--osinfo-mongo--prod
+  - name: rj-cvl--osinfo-pdf-reconstruct--prod
     version: "{{ get-commit-hash.stdout }}"
-    entrypoint: pipelines/rj_cvl__osinfo_mongo/flow.py:rj_cvl__osinfo_mongo
+    entrypoint: pipelines/rj_cvl__osinfo_pdf_reconstruct/flow.py:rj_cvl__osinfo_pdf_reconstruct
     work_pool:
       name: k3s-pool
       work_queue_name: default
       job_variables:
         image: "{{ build-image.image_name }}:{{ build-image.tag }}"
-        command: uv run --package rj_cvl__osinfo_mongo -- prefect flow-run execute
+        command: uv run --package rj_cvl__osinfo_pdf_reconstruct -- prefect flow-run execute
         secretName: prefect-jobs-secrets
         image_pull_policy: Always
